This copyright registration letter
The copyright of this programming was officially registered on May 31, 2021, by the Directorate General of Intellectual Property—Ministry of Law and Human Rights, the Republic of Indonesia, with the registration number [EC00202125523] nd [000252427], valid for 50 (fifty) years from the first announcement of the work. This copyright registration letter or related rights products are in accordance with article 72 of law number 28 of 2014 concerning copyright.